+++ socket/client/src/backends/wasm_bindgen/mod.rs
@@ -1,11 +1,12 @@
 mod addr_cell;
-mod webrtc_internal;
+mod peer_connection;
+mod data_channel;
 
 mod packet_receiver;
 mod packet_sender;
 mod socket;
 
-pub use webrtc_internal::PeerConnection;
+pub use peer_connection::PeerConnection;
 pub use packet_receiver::PacketReceiverImpl;
 pub use packet_sender::PacketSender;
 pub use socket::Socket;
